Backend Code Complete

+++ server/src/main/java/com/skillconnect/server/service/UserService.java
@@ -9,15 +9,15 @@ import java.util.Optional;
 public interface UserService {
     
     User saveUser(User user);
-    Optional<User> findById(Long userId);
+    Optional<User> findById(int userId);
     User findByEmail(String email);
     List<User> findAllUsers();
     User updateUser(User user);
-    void deleteUser(Long userId);
+    void deleteUser(int userId);
     boolean existsByEmail(String email);
     boolean existsByUsername(String username);
-    List<User> getFollowers(Long userId);
-    List<User> getFollowing(Long userId);
-    User updateProfile(Long userId, String firstName, String lastName, String bio, String profileImage);
-    boolean changePassword(Long userId, String currentPassword, String newPassword);
+    List<User> getFollowers(int userId);
+    List<User> getFollowing(int userId);
+    User updateProfile(int userId, String firstName, String lastName, String bio, String profileImage);
+    boolean changePassword(int userId, String currentPassword, String newPassword);
 }
